requirements.txt
langgraph>=0.4,<1.0
langchain-openai>=0.3,<1.0
python-dotenv>=1.0,<2.0
